Use new farmLog implementation.
